% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{Estep}
\alias{Estep}
\title{Perform the E-step in the EM algorithm}
\usage{
Estep(H, Ic, R1, dead)
}
\arguments{
\item{H}{Hazards per individual (in columns)}

\item{Ic}{Censoring interval per individual, coded as 0/1 (in columns)}

\item{R1}{Left truncation interval per individual, coded as 0/1 (in columns)}

\item{dead}{Boolean vector (TRUE is event, FALSE is right censored)}
}
\value{
A list with two matrices
\item{Y}{Expected probability of event per bin per subject}
\item{R}{Expected probability of at risk per bin per subject}
}
\description{
Perform the E-step in the EM algorithm
}
