% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillplot.R
\name{fillplot}
\alias{fillplot}
\title{Fills space between two lines in a graph}
\usage{
fillplot(x, y1, y2, col)
}
\arguments{
\item{x}{Points on the x-axis}

\item{y1}{First set of points on y-axis}

\item{y2}{Second set of points on y-axis}

\item{col}{The color to fill space with}
}
\value{
Nothing
}
\description{
Taken from mstate
}
