% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ova.R
\docType{data}
\name{Ova}
\alias{Ova}
\title{Ovarian cancer data}
\format{
A dataframe with five columns:

\describe{
  \item{\code{Diameter}}{}
  \item{\code{FIGO}}{}
  \item{\code{Karnofsky}}{}
  \item{\code{time}}{}
  \item{\code{d}}{death}
}
}
\source{
tba
}
\usage{
data(Ova)
}
\description{
Ovarian cancer data
}
\keyword{datasets}
