% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_icfit.R
\name{print.icfit}
\alias{print.icfit}
\title{Print method for an object of class `icfit`}
\usage{
\method{print}{icfit}(x, digits = max(1L, getOption("digits") - 3L), alpha = 0.05, ...)
}
\arguments{
\item{x}{The object of class 'icfit' to be printed}

\item{digits}{Number of digits to be printed}

\item{alpha}{Alpha level to be used of confidence interval ((1-alpha) * 100 percent)}

\item{\dots}{Further arguments to print}
}
\value{
No return value
}
\description{
Print method for an object of class `icfit`
}
